Added setElementCollisionsEnabled and getElementCollisionsEnabled server-side, including a map property for the collisions. Also added getElementCollisionsEnabled client-side.
<object model="1337" posX="-709" posY="962" posZ="12" scale="0.1" collisions="false"/>
<object model="1337" posX="-714" posY="962" posZ="12" scale="0.1" collisions="true"/>

Spawns 2 miniature garbage bins near the "Play" spawn point. One of them will have collisions, the other won't.

+++ MTA10_Server/mods/deathmatch/logic/lua/CLuaFunctionDefinitions.cpp
@@ -6845,7 +6845,7 @@ int CLuaFunctionDefinitions::SetObjectScale ( lua_State* luaVM )
         CObject* pObject = lua_toobject ( luaVM, 1 );
         if ( pObject && IS_OBJECT ( pObject ) )
         {
-            float fScale = lua_tonumber ( luaVM, 2 );
+            float fScale = (float) lua_tonumber ( luaVM, 2 );
 
             if ( CStaticFunctionDefinitions::SetObjectScale ( pObject, fScale ) )
             {
